Regex pattern support closes #1
This commit adds support for creating of regex pattern for student numbers.
Created pattern can match student numbers that are in range of provided argumnets.
The output is a reglar expression pattern that can be used to match certain studeny numbers in string.

Pattern created would sometimes match numbers that are not student numbers but in rare cases.
There are less chances for that to happen and in most cases can be ignored.
Created pattern is so strict on what should be matched and would filter-out most non-matching student numbers.

This commit has modifications to analyse and generate module which also involved refactoring.

Changes
-------
analyse.py - Constants from generate.py were added and also new ones added.
           - Validation of year and position was updated.
           - calculate_year_capacity() function was added.
           - get_capacity_from_kwargs() and get_first_position_from_kwargs() were added.

generate.py - Constants removed to analyse.py.
            - Added helper functions to help in raising exceptions and getting keywords argumnets.
            - Added functions for creating regex pattern for position part, year part and whole student number.

Note: Generated pattern was never even tested if it works on real string.
    : Its possibe that it does not even work.