Fix regex pattern closes #3
Feature added from previous commit and issue #1 had bugs as the regex
pattern was not working as expected. It was caused by misuse and misuderstanding
of regex characters classes.

It was based on feature that regular expressions do not support. Regex is used for
maching pattern based on pattern maching characters in string. Matching a number would
require maching the characters not the value of the number as in python.

This commit added changes to fix that and make regex pattern work even if not as intended
to be.

See issues #1, #3 for more info.