Update: access.py, analyse.py, generate.py
This commit was intended for adding support for reading student numbers from
files and also filtering them based on wheather they are valid. To make that
possible, it required that changes be made to ensure the were no exceptions raised.

Bugs within analyse.py that caused exceptions when using int() were fixed.
That was bacause sometimes year or position part can be empty or not be a
digit which causes exception when using int().

student_number_allowed() was added to generate.py as extension to
analyse.student_number_valid() function. It is special in that it support extra
argumnet specific to generate.py such as 'start_pos' which is not available on
the later.

Functions not indended to be used access.py were made private. filter_student_numbers()
function was added for filtering student number by their validity based on provided
keywords optional arguments.

extract_student_numbers() was updated to use filter_student_numbers() to avoid returning
student numbers that are not valid or not expected since regex for matching student numbers
is not accurate.

extract_student_numbers_file() was also added to extactiong student numbers from files. It
may use navaly if file is not plain text file. Pdf, docx, etc files are supported but navaly will
need to be imported manually for them to be supported.
analyse.py had recived bug fixes regarding functions involving validating